srcnnimg/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
Pillow==8.2.0
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
